ct_enter(): enter mnt namespace last
mnt needs to be set last since it changes the filesystem.

Runnig vzctl start on upstream kernel fails for me on CentOS 6.5.
I used to run vzkernel but reverted to stock to be able to run
docker / lxc at the same time.

With strace I noticed it's failing to find /proc/<pid>/ns/pid and
I traced that to the loop in hooks_ct.c. After it does setns
on the mnt file the pid one no longer can be opened, even though
it's reported by readdir.

The attached patch fixes this by skipping mnt in the loop
and doing it explicitly afterwards.

https://bugzilla.openvz.org/3038