vzctl snapshot-switch: add --must-restore
It has been pointed out to me that snapshot-switch ignores the failure
to restore a CT from a dump file. Also, not every snapshot contains a
dump file, so it's not clear whether we will have a running or stopped
container after this operation.

Introduce --must-restore flag to eliminate such uncertainty -- if it is
set, a dump file is required (i.e. its absense is treated as an error
rather than ignored), and a sucessful restore/resume is also required
(i.e. the failure to restore is treated as an error rather than warning).

Man page modified accordingly.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>