vzctl.spec: require recent RHEL6 kernel
Because of commit bb1956e vzctl can't be used on kernel older than
042stab078.22 as those kernels do not set unset UBCs to unlimited,
resulting in an inability to start a container.

We could potentially fix it by setting unset UBCs to unlimited only
on CT start. But as this kernel is rather old we'd better just require
a more recent kernel.

Due to the fact that vzctl also works on RHEL5 kernel, we use
"Conflicts:" rather than "Requires:".

https://bugzilla.openvz.org/3094

Reported-by: Avi Brender <abrender@elitehosts.com>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>