vzmigrate: random ports for ploop copy
This was on my todo list, and also requested via bugzilla.

We generate a random port number (in the range 1024-32767 as lower
ports are "privileged" and upper are "ephemeral") and try to run
ssh port forwarding command. If it fails, retry with a different
port number, up to 5 times. If it still fails with 5 different
random port numbers, we assume it is not going to work and fallback
the non-feedback ploop copy (same as it was done before).

https://bugzilla.openvz.org/3052

Reported-by: Pavel Odintsov <pavel.odintsov@gmail.com>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>