vzmigrate: use DUMPDIR for CT dump
As pointed out in bug #3054, sometimes CT dump file may be very big
so it might not fit in /var/tmp. Chances are high though that /vz
do have enough free disk space for it, so let's use /vz/dump
(or whatever DUMPDIR is set to).

Note this patch also handles the case when DUMPDIR is different on the
destination server.

So, we only use /var/tmp for quota dump which should be pretty small.

https://bugzilla.openvz.org/3054

Reported-by: Pavel Odintsov <pavel.odintsov@gmail.com>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>