add_dev_param(): don't modify second argument
Commit ea64ee6 has a bad side effect: after calling merge_dev()
the device names in vps_p list (which is supposed to have configuration
as read from ve.conf) disappear. This is caused by the hunk modifying
add_dev_param().

This patch fixes that side effect, also marking the second argument
const so this won't happen in the future.

This is not a bug fix per se (as no one was using dev in vps_p list),
but is a requirement for the next commit.

Based on a patch by Andrey Vagin <avagin@openvz.org>:
  https://bugzilla.openvz.org/3186#c3

Signed-off-by: Kir Kolyshkin <kir@openvz.org>