vzmigrate: check CPU caps for suspended CT
This commit adds a check for CPU capabilities when migrating a
suspended container. Such a check is needed because if CPU caps
differ, a container might not survive restore after migrating.

To implement this check, a support from the kernel is required
(CPT_TEST_CAPS ioctl for /dev/cpt). Such support was errorneously
removed from the kernel 042stab101.5, but then restored in 042stab107.1,
so any kernels in between those return EAGAIN which we treat as OK
since we can't perform the check needed. In such case, though,
we issue a warning that the kernel should be updated.

vzcptcheck modified accordingly, together with usage and man page.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>