vzctl.spec: drop the Conflicts: vzkernel
The change this patch is reverting was introduced to fix bug #3094.
Unfortunately, it breaks the update for RHEL5-based OpenVZ kernels,
which is the subject of bug #3219.

The only solution I found is to list all the old RHEL6-based kernel
explicitly. It would be good enough, but looks ugly.

As the newest RHEL6-based kernel conflicting with this vzctl is at least
2 years old, let's just drop the "Conflicts: vzkernel" line. This makes
vzctl to fail on older kernels, but well, two-years old kernels is full
of securuty holes and bugs, please, just upgrade it already.

https://bugzilla.openvz.org/3219

Reported-by: М. Alhimenko <lithium@mail.ru>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>