vps-create.sh: use stat -f instead of df
On a node with a stale NFS mount (not related to /vz), vzctl create
fails after a long timeout. It happens because vps-create calls
df $VE_ROOT, but for some odd reason df wants to stat() all the
mount points and it hangs for a while and then fails on that stale mount.

The solution is to use stat -f rather than df.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>