vzoversell: handle unlimited RAM CTs
Some containers might be configured as unlimited
(i.e. infinite physpages and privvmpages limit).

Also, a container might appear as unlimited while
it is being restored from dump (as beancounter limits
are initially set to unlimited and only set to proper
values at the end of the restoration process).

In case such unlimited containers are present,
overcommitment figures shown doesn't make sense,
for example:

> RAM        available: 59.4G  allocated:   16Z  oversell: 29630589707078%
> Swap       available:  177G  allocated: 18.9G  oversell: 10%
> RAM+Swap   available:  237G  allocated:   16Z  oversell: 7411239850072%

The fix is to skip such containers, as we already do for
swap.

Notes:
1 we do not just detect the unlimited container, but also
  mark it with a flag ("U") and explain why.
2 RAM usage and peak usage of an unlimited container
  is still being included into totals, it's just the limit
  that is being ignored.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>