ve_is_mounted(): describe the ploop semi-mounted case
I know it looks weird, but this is it -- originally I wanted
to modify vps_is_mounted() is it would return "mounted" in case
ploop device is mounted. But it would be a wrong thing to do
since if CT status is "mounted", CT files are expected to be
available from VE_PRIVATE, and it's not the case.

So, the code is all here but it's commented out.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>